17 September 2023
Given a String s, In one operation we can delete any occurences of "geek" as a substring in the String. Find the number of delete operations you have to perform to make the String Empty. 

Example 1:

Input:
s = "gegeekek"
Output:
2
Explanation:
gegeekek
We can first delete the bold characters
which form "geek", and then delete the
remaning string "geek" again from the string.
So, we will need 2 operations

Example 2:

Input:
s = "geekkgee"
Output:
-1
Explanation:
It's not possible to make the String empty.
Your Task:
You don't need to read input or print anything. Your task is to complete the function makeStringEmpty() which takes a String s as input and returns the minimum number of operations required to delete the String entirely. If it's not possible to do so return -1.

Constraints:
1 <= |s| <= 10^5