17 September 2023
Given an integer D and an array A[ ] of N integers. At each operation you can select the elements which are at index i and (i + D)%N and you can swap the both elements. The task is to check if it is possible to sort the array by performing the above operation zero or more times.

Note: 0 based indexing is used.

Example 1:

Input:
N = 4, D = 2
A[] = {3, 2, 1, 5}
Output: 
Yes
Explanation: 
We can sort the array in one operation.
1st operation: swap(A[i], A[(i+D)%N])
where i = 0.
Example 2:

Input:
N = 2, D = 0
A[] = {5, 2}
Output: 
No 
Explanation: 
It is not possible to sort the array.
Your Task: 
You don't need to read input or print anything. Complete the function check( ) which takes integer D, integer N and array A[ ] as input parameters and returns true if it is possible to sort the array by performing the above operation zero or more times else return false.

Constraints:
1 ≤ N ≤ 10^5
1 ≤ A[i] ≤ N
0 ≤ D < N