10 September 2023
You are given a string S of lowercase Latin characters with size N. Your task is to find the lexicographically smallest substring with the maximum frequency.

Note: String p is lexicographically smaller than string q, if p is a prefix of q, is not equal to q or there exists i, such that pi < qi and for all j < i it is satisfied that pj = qj. For example, abc is lexicographically smaller than abcd, abd is lexicographically smaller than abec, afa is not lexicographically smaller than ab and a is not lexicographically smaller than a.

Input:
N = 4
S = "gfgg"
Output: 
"g"
Explanation: 
Substring "g" is present 3
in the string and it can be proved
that it is the lexicographically 
smallest one
Example 2:

Input:
N = 1
S = "a"
Output: 
"a"
Explanation: 
The only possible 
substring is "a"
Your Task:  
You don't need to read, input, or print anything. Your task is to complete the function solve( ), which takes N and string S as input parameters and returns the answer to the problem.

 

Constraints:
1 ≤ N ≤ 10^5
String S consists of lowercase Latin characters