03 September 2023
Geek is fond of explosives. He has lined up a row of N boxes containing explosives in his garden. The boxes are of K different colours. The ith box is of colour col[i]. 
If three or more boxes of similar colour occur together, all of them will explode without affecting their neighboring boxes which are not of the same colour.
In the initial arrangement of the boxes, no three boxes of the same colour will occur together. Geek has an extra box with him of colour C. He can insert that box between any two boxes in the row to set off the explosion. Help geek find the maximum number of boxes can be explode. 

Note: The extra box inserted by Geek is not counted in the final answer.


Example 1:

Input:
N = 6, K = 2, C = 2
col[] = {1, 1, 2, 2, 1, 1}
Output: 6
Explanation: Insert the extra box between 
the 3rd and 4th box. Initially it will 
explode all the boxes of colour 2 and 
then all the boxes of colour 1.

Example 2:

Input:
N = 1, K = 1, C = 1
col[] = {1}
Output: 0
Explanation: No group of 3 or more 
boxes are possible

Your Task:  
You don't need to read input or print anything. Your task is to complete the function maxBoxes() which takes the integer N, K, C, and col[ ] as input parameters and returns the maximum number of boxes that can be exploded.
 

Constraints:
1 < N, K < 10^3
1 < C < K
1 < col[i] < K